requirements/base.txt
Django==1.6.5
Pillow==2.3.0
South==0.8.4
Unipath==1.0
aldryn-blog==0.3.12
cmsplugin-filer==0.9.5
django-cms==3.0.2
django-compressor==1.4
django-mathfilters==0.1.3
django-reversion==1.8.0
django-signup==0.3.1
djangocms-text-ckeditor==2.1.4
pathlib==0.97
pytz==2014.4
solid-i18n==0.2.1
